ISIS-2877: let composite-value-mixin names correspond to qualifier names
- in support of per feature customization
